renaming strainer to strain, keeping file name the same as the utility it defines
