+++ src/strain.ts
